# docs/openapi.yaml
openapi: 3.0.0
info:
  title: ArchKit Unified API
  description: OpenAPI documentation for both Monolith and Hybrid architectures.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Monolith Server
  - url: http://localhost:3001
    description: Hybrid Product Service
  - url: http://localhost:3002
    description: Hybrid Sales Service
  - url: http://localhost:3003
    description: Hybrid Inventory Service
tags:
  - name: Product
    description: Product management operations
  - name: Sales
    description: Sales transaction processing
  - name: Inventory
    description: Inventory and stock management
  - name: Diagnostics
    description: Utility and health endpoints
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: "Validation failed"
        error:
          type: string
          example: "Bad Request"
        errors:
          type: array
          items:
            type: string
          example: ["name must be a string", "price must be positive"]
      required:
        - statusCode
        - message
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          type: number
          minimum: 0.01
      required:
        - id
        - name
        - price
    CreateProductDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          type: number
          minimum: 0.01
      required:
        - name
        - price
    UpdateProductDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        price:
          type: number
          minimum: 0.01
    SalesItem:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
      required:
        - productId
        - quantity
    SalesTransaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesItem'
          minItems: 1
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - items
        - totalAmount
        - createdAt
    CreateTransactionDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesItem'
          minItems: 1
      required:
        - items
    InventoryBalance:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        balance:
          type: integer
      required:
        - productId
        - balance
    AdjustStockDto:
      type: object
      properties:
        productId:
          type: string
          format: uuid
        delta:
          type: integer
      required:
        - productId
        - delta
paths:
  /products:
    get:
      tags:
        - Product
      summary: Get all products
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags:
        - Product
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductDto'
      responses:
        '201':
          description: The product was successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      tags:
        - Product
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A product object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
        - Product
      summary: Update a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductDto'
      responses:
        '200':
          description: The product was successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags:
        - Product
      summary: Delete a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: The product was successfully deleted
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sales/transaction:
    post:
      tags:
        - Sales
      summary: Process a sales transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionDto'
      responses:
        '201':
          description: Transaction processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesTransaction'
        '400':
          description: Validation error or insufficient stock
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /sales/transactions/{id}:
    get:
      tags:
        - Sales
      summary: Get a sales transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A sales transaction object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesTransaction'
        '404':
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/adjust:
    post:
      tags:
        - Inventory
      summary: Adjust stock level for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustStockDto'
      responses:
        '201':
          description: Stock adjusted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryBalance'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /inventory/{productId}:
    get:
      tags:
        - Inventory
      summary: Get inventory balance for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current inventory balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryBalance'
  /health:
    get:
      tags:
        - Diagnostics
      summary: Health check
      responses:
        '200':
          description: Service is healthy
  /diagnostics/ping:
    get:
      tags:
        - Diagnostics
      summary: Ping the service
      responses:
        '200':
          description: Pong